fix: migrate from Qwik to Solidjs, update bun dependencies
