+++ astro.config.mjs
@@ -1,7 +1,6 @@
 import { defineConfig } from "astro/config";
 import mdx from "@astrojs/mdx";
 import sitemap from "@astrojs/sitemap";
-import qwikdev from "@qwikdev/astro";
 import tailwindcss from "@tailwindcss/vite";
 import solidJs from "@astrojs/solid-js";
 import netlify from "@astrojs/netlify";
@@ -15,9 +14,6 @@ export default defineConfig({
   integrations: [
     mdx(),
     sitemap(),
-    qwikdev({
-      include: "**/qwik/*",
-    }),
     solidJs({
       devtools: true,
       include: ["**/solid/*", "**/node_modules/@suid/material/**"],
