--- src/components/qwik/AnimatingDots.tsx
@@ -1,42 +1,0 @@
-/** @jsxImportSource @builder.io/qwik */
-import { component$, useVisibleTask$ } from "@builder.io/qwik";
-import gsap from "gsap";
-
-export default component$(() => {
-  useVisibleTask$(() => {
-    let blockAnimationTimeline = gsap.timeline({
-      delay: 0.01,
-    });
-
-    let blocks = document.querySelectorAll("#blocks > div");
-
-    blockAnimationTimeline
-      .set(blocks, {
-        opacity: 0,
-        xPercent: -10,
-      })
-      .add("blockLoad")
-      .to(
-        blocks,
-        {
-          opacity: 1,
-          xPercent: 0,
-          duration: 0.35,
-          delay: 0.25,
-          stagger: 0.15,
-        },
-        "blockLoad",
-      );
-  });
-
-  return (
-    <div id="blocks" class="flex justify-center gap-6 md:gap-10 relative" inert>
-      <div class="block w-[7px] h-[7px] md:w-[10px] md:h-[10px] bg-terminal-green opacity-0"></div>
-      <div class="block w-[7px] h-[7px] md:w-[10px] md:h-[10px] bg-terminal-light-blue opacity-0"></div>
-      <div class="block w-[7px] h-[7px] md:w-[10px] md:h-[10px] bg-terminal-pale-purple opacity-0"></div>
-      <div class="block w-[7px] h-[7px] md:w-[10px] md:h-[10px] bg-terminal-purple opacity-0"></div>
-      <div class="block w-[7px] h-[7px] md:w-[10px] md:h-[10px] bg-terminal-orange opacity-0"></div>
-      <div class="block w-[7px] h-[7px] md:w-[10px] md:h-[10px] bg-terminal-red opacity-0"></div>
-    </div>
-  );
-});
